// src/services/route53zone/config-types.ts
/*
 * Copyright 2017 Brigham Young University
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 */
import { ServiceConfig, Tags } from '../../datatypes/index';

export interface Route53ZoneServiceConfig extends ServiceConfig {
    name: string;
    private?: boolean;
}

export interface HandlebarsRoute53ZoneTemplate {
    name: string;
    tags: Tags;
    vpcs?: HandlebarsRoute53ZoneVpc[];
}

export interface HandlebarsRoute53ZoneVpc {
    id: string;
    region: string;
}